darwin: Improve JAVA_HOME assumptions
In the case of installing oracle's jdk package, jdk home may be in
a path such as /Library/Java/JavaVirtualMachines/jdk1.7.0_51.jdk/Home.
In this case, /usr/libexec/java_home reports the correct path even when
$JAVA_HOME is unset.

If JAVA_HOME is unset on darwin, check for the existence of the
/usr/libexec/java_home utility. If it points to the requested java major
version, then use it to set JAVA_HOME. Otherwise, set the fallback path.

Change-Id: Id698078d2fc8d6d258f87ebee211f49882c4dc56