darwin: Improve JAVA_HOME assumptions
Summary:
In the case of installing oracle's jdk package, jdk home may be in
a path such as /Library/Java/JavaVirtualMachines/jdk1.7.0_51.jdk/Home.
In this case, /usr/libexec/java_home reports the correct path even when
$JAVA_HOME is unset.

If JAVA_HOME is not set on darwin, check for the existence of the
/usr/libexec/java_home utility, then use it to set JAVA_HOME before
setting the standard osx java 1.6 path.

Test Plan: gtg

Reviewers: raymanfx, randall.rushing

Reviewed By: raymanfx

Differential Revision: http://review.codefirex.com/D465